requirements.txt
gunicorn
Django
djangorestframework
django-cors-headers
drf_spectacular
whitenoise